clippy on sim
